remove vm/streaming
