+++ config/config.go
@@ -32,7 +32,6 @@ func (c *Config) GetMempoolSize() int                    { return 2_048 }
 func (c *Config) GetMempoolPayerSize() int               { return 32 }
 func (c *Config) GetMempoolExemptPayers() [][]byte       { return nil }
 func (c *Config) GetMempoolVerifyBalances() bool         { return true }
-func (c *Config) GetStreamingPort() uint16               { return 0 } // auto-assigned
 func (c *Config) GetStreamingBacklogSize() int           { return 1024 }
 func (c *Config) GetStateHistoryLength() int             { return 256 }
 func (c *Config) GetStateCacheSize() int                 { return 65_536 } // nodes
